Fix TypeScript error in cleanDb.ts

+++ server/src/seeds/cleanDb.ts
@@ -1,14 +1,13 @@
-import models from '../models/index.js';
-import db from '../config/connection.js';
-
 export default async (modelName: "Question", collectionName: string) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
-      name: collectionName
-    }).toArray()
+    if (models[modelName]) {
+      let modelExists = await models[modelName].db.db.listCollections({
+        name: collectionName
+      }).toArray()
 
-    if (modelExists.length) {
-      await db.dropCollection(collectionName);
+      if (modelExists.length) {
+        await db.dropCollection(collectionName);
+      }
     }
   } catch (err) {
     throw err;
